Migration files updated to keep up postgre syntax

+++ src/main/resources/db/migration/migration/V8__create_invoices_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS invoices (
     id                  SERIAL PRIMARY KEY,
     account_id          INTEGER NOT NULL REFERENCES accounts(id) ON DELETE CASCADE,
-    electricity_units   DOUBLE NOT NULL,
+    electricity_units   DECIMAL NOT NULL,
     invoice_date        DATE NOT NUll
 )
